+++ src/st/np3/e_init.c
@@ -200,8 +200,8 @@ u16 D_80180AB4[] = {0x8003, 0x0001, 0x0048, 0x028C, 0x001D};
 u16 D_80180AC0[] = {0x8003, 0x0000, 0x0048, 0x028C, 0x001E};
 u16 D_80180ACC[] = {0x8004, 0x0001, 0x004A, 0x02B2, 0x001B};
 u16 D_80180AD8[] = {0x8004, 0x0000, 0x004A, 0x02B2, 0x001C};
-u16 D_80180AE4[] = {0x8002, 0x0001, 0x0048, 0x02C5, 0x0069};
-u16 D_80180AF0[] = {0x8002, 0x0001, 0x0048, 0x02C5, 0x0002};
+u16 g_BoneScimitarInit[] = {0x8002, 0x0001, 0x0048, 0x02C5, 0x0069};
+u16 g_ScimitarPartsInit[] = {0x8002, 0x0001, 0x0048, 0x02C5, 0x0002};
 u16 D_80180AFC[] = {0x8005, 0x0000, 0x004C, 0x02C8, 0x0040};
 u16 D_80180B08[] = {0x8007, 0x0000, 0x004D, 0x02D3, 0x006E};
 u16 D_80180B14[] = {0x8008, 0x0010, 0x0049, 0x02CB, 0x0014};
